+++ src/form/hooks/useForm.ts
@@ -90,7 +90,7 @@ class FormStore {
     // No need to cost perf when nothing need to watch
     if (this.watchList.length) {
       // @ts-ignore
-      const values = this.getFieldsValue?.(namePath);
+      const values = this.getFieldsValue?.([namePath]);
 
       this.watchList.forEach((callback) => {
         callback(values, namePath);